(#25)  REFACT Data models to model/data
