+++ mywifipass/app/src/main/java/com/example/mywifipass/model/data/LoginCredentials.kt
@@ -1,0 +1,11 @@
+package app.mywifipass.model.data
+
+data class LoginCredentials(
+    var url: String = "",
+    var login: String = "",
+    var pwd: String = "",
+) {
+    fun isNotEmpty(): Boolean {
+        return login.isNotEmpty() && pwd.isNotEmpty() && url.isNotEmpty()
+    }
+}
