+++ mywifipass/app/src/main/java/com/example/mywifipass/ui/Cards.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
-import app.mywifipass.backend.database.Network
+import app.mywifipass.model.data.Network
